update the plugins and imports
